Extract function.ts
